Make the Robots/Addresses/Requests page header more compact
