+++ scss/page-header.scss
@@ -8,9 +8,16 @@ $big-text: 24px;
   background-color: $navbar-bg;
   color: white;
   position: relative;
+  height: 200px;
   margin-bottom: 20px;
-  height: 240px;
   margin-bottom: 0;
+  flex-shrink: 0;
+  display: flex;
+  flex-direction: column;
+}
+
+.page-header--big {
+  height: 240px;
 }
 
 .page-header__header {
@@ -30,6 +37,11 @@ $big-text: 24px;
 }
 
 .page-header__body {
+  flex-grow: 1;
+  display: flex;
+  .Row {
+    flex-grow: 1;
+  }
 }
 
 .page-header__filters {
