+++ scss/sessions/session_toolbar.scss
@@ -1,0 +1,73 @@
+@import '../vars';
+@import '../mixins';
+
+$checkbox-size: 20px;
+
+.session-toolbar {
+  flex-grow: 1;
+  display: flex;
+  align-items: center;
+  font-family: $default-light-font;
+}
+
+.session-toolbar__checkbox {
+  display: inline-flex;
+  cursor: pointer;
+  font-size: 20px;
+  align-items: center;
+}
+
+.session-toolbar__checkbox-robots__text {
+  text-transform: capitalize;
+}
+
+.session-toolbar__checkbox__checkbox {
+  position: relative;
+  display: inline-block;
+  border-radius: 3px;
+  height: $checkbox-size;
+  width: $checkbox-size;
+  border-style: solid;
+  border-width: 2px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+  margin-right: 5px;
+  input {
+    display: none;
+  }
+
+  @each $status, $color in $statuses {
+    &--#{$status} {
+      border-color: $color;
+    }
+  }
+}
+
+.session-toolbar__checkbox__checkbox--none-checked {
+  @each $status, $color in $statuses {
+    &.session-toolbar__checkbox__checkbox--#{$status} {
+      background-color: $color;
+    }
+  }
+}
+
+.session-toolbar__checkbox__checkbox--checked {
+  @each $status, $color in $statuses {
+    &.session-toolbar__checkbox__checkbox--#{$status} {
+      background-color: $color;
+    }
+  }
+
+  &::before {
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    content: '';
+    fill: $brand-purple;
+    background-image: url(../../assets/checkmark-white.svg);
+    @include white-to-brand-purple();
+    background-size: 100%;
+  }
+}
