+++ src/react/pages/addresses.jsx
@@ -9,14 +9,9 @@ import {
   reputationResolver,
 } from '../address/address_resolvers';
 
-import RobotsRowHeader from '../robots/robots_row_header';
+import SessionToolbar from '../sessions/session_toolbar';
 import Sessions from '../sessions/sessions';
-import SessionsVisualisationSwitch from '../sessions/sessions_visualisation_switch';
-import {
-  routePropType,
-  addressSessionsPropType,
-  robotsMetricsPropType,
-} from '../prop_types';
+import { routePropType, addressSessionsPropType } from '../prop_types';
 
 import '../../../scss/sessions/sessions_page.scss';
 import '../../../scss/addresses_page.scss';
@@ -26,7 +21,7 @@ const rowClassResolver = address => {
   return status ? `addresses__table__row--${status}` : '';
 };
 
-const AddressesPage = ({ route, addresses, robotsMetrics }) => (
+const AddressesPage = ({ route, addresses }) => (
   <div className="addresses-page page--sessions">
     <div className="page-header page-header--addresses">
       <div className="page-header__header">
@@ -36,12 +31,9 @@ const AddressesPage = ({ route, addresses, robotsMetrics }) => (
               Top Addresses {`(${timeDisplay()})`}
             </span>
           </Col>
-          <Col md="50%">
-            <SessionsVisualisationSwitch route={route} />
-          </Col>
         </Row>
       </div>
-      <RobotsRowHeader metrics={robotsMetrics} route={route} />
+      <SessionToolbar route={route} />
     </div>
     <Sessions
       sessions={addresses}
@@ -60,7 +52,6 @@ const AddressesPage = ({ route, addresses, robotsMetrics }) => (
 AddressesPage.propTypes = {
   route: routePropType.isRequired,
   addresses: addressSessionsPropType.isRequired,
-  robotsMetrics: robotsMetricsPropType.isRequired,
 };
 /* eslint-enable react/no-typos */
